+++ src/server/Account/AccountServer.ts
@@ -1,0 +1,13 @@
+import { ApiModule } from "../ApiModule";
+import { iNode } from "../iNode";
+import { Account } from "../../crosscommon/entities/Account";
+
+export class AccountServer {
+    list = (node: iNode) => {
+        let api: ApiModule = new ApiModule(new Account());
+
+        api.list(node).then((response) => {
+            node.response.end(JSON.stringify(response));
+        });
+    };
+}