+++ src/server/Place/PlaceRoute.ts
@@ -1,0 +1,16 @@
+import * as express from 'express';
+import { iNode } from "../iNode";
+import { PlaceServer } from './PlaceServer';
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    let server: PlaceServer = new PlaceServer();
+    let node: iNode = {
+        request: req
+        , response: res
+    };
+    server.list(node);
+});
+
+export { router };